% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpidemiologyModelModifications.R
\name{AddModelIdentifier}
\alias{AddModelIdentifier}
\title{A model object check.}
\usage{
AddModelIdentifier(model, id)
}
\arguments{
\item{model}{A model object.}

\item{id}{A number or string to used as ID.}
}
\value{
A model.
}
\description{
Adds a specified identifier id to all stocks and rates on the model m.
}
\concept{Epidemiology Model Modifiers}
