% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpidemiologyModelModifications.R
\name{AddModelRates}
\alias{AddModelRates}
\title{Add model rates.}
\usage{
AddModelRates(
  model,
  rateDescriptions,
  rateValues,
  replaceExistingRatesQ = TRUE
)
}
\arguments{
\item{model}{A model object.}

\item{rateDescriptions}{A character vector with named elements.}

\item{rateValues}{A numerical vector with named elements.}

\item{replaceExistingRatesQ}{Should existing rates be replaced or not?}
}
\value{
A model.
}
\description{
Adds rates to model object.
}
\details{
This function is useful when working with rate functions.
If a rate in the model object is changed to be a function (instead of number)
then in order to do batch simulations it is convenient to have the parameters
of that function as rates in the model object.
}
\seealso{
Other Epidemiology Model Modifier functions: 
\code{\link{AddModelIdentifier}()},
\code{\link{JoinModels}()},
\code{\link{MakeMigrationEquations}()},
\code{\link{TakeRHSFunctionBodyLines}()},
\code{\link{ToSiteCompartmentalModel}()}
}
\concept{Epidemiology Model Modifier functions}
