% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECMMon.R
\name{ECMMonAddRates}
\alias{ECMMonAddRates}
\title{Add model rates.}
\usage{
ECMMonAddRates(
  ecmObj,
  rateDescriptions,
  rateValues,
  replaceExistingRatesQ = TRUE
)
}
\arguments{
\item{ecmObj}{An ECMMon object.}

\item{rateDescriptions}{A character vector with named elements.}

\item{rateValues}{A numerical vector with named elements.}

\item{replaceExistingRatesQ}{Should existing rates be replaced or not?}
}
\value{
An ECMMon object.
}
\description{
Adds specified rates to monad's default model.
}
\details{
This function is useful when working with rate functions.
If a rate in monad's model object is changed to be a function (instead of number)
then in order to do batch simulations it is convenient to have the parameters
of that function as rates in the model object.
}
\seealso{
Other Simulation functions: 
\code{\link{ECMMonAssignInitialConditions}()},
\code{\link{ECMMonAssignRateValues}()},
\code{\link{ECMMonBatchSimulate}()},
\code{\link{ECMMonExtendByAdjacencyMatrix}()},
\code{\link{ECMMonSimulate}()}
}
\concept{Simulation functions}
