% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECMMon.R
\name{ECMMonAssignInitialConditions}
\alias{ECMMonAssignInitialConditions}
\title{Assign initial conditions to model.}
\usage{
ECMMonAssignInitialConditions(ecmObj, initConds)
}
\arguments{
\item{ecmObj}{An ECMMon object.}

\item{initConds}{A numerical vector with named elements.
The names are expected to be known stocks in the model.}
}
\value{
An ECMMon object.
}
\description{
If there is a multi-site model assign initial conditions to it.
Otherwise the initial conditions are assigned to the single-site model.
}
\seealso{
Other Simulation functions: 
\code{\link{ECMMonAssignRateValues}()},
\code{\link{ECMMonExtendByAdjacencyMatrix}()},
\code{\link{ECMMonSimulate}()}
}
\concept{Simulation functions}
