% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECMMon.R
\name{ECMMonBatchSimulate}
\alias{ECMMonBatchSimulate}
\title{Batch simulate.}
\usage{
ECMMonBatchSimulate(ecmObj, params, maxTime, resultForm = "simple", ...)
}
\arguments{
\item{ecmObj}{An ECMMon object.}

\item{params}{A data frame of a list of numeric vectors.
If \code{params} is a list of numeric vectors then \code{params} is expected to have named elements,
and \code{names(params)} to be known parameter names.}

\item{maxTime}{A numerical non-negative value for the maximum simulation time.}

\item{resultForm}{A string or NULL.
If a string it is expected to be one of "list", "simple", or "export".
NULL is the same as "simple".}

\item{...}{Additional parameters of \code{\link{ECMMonSimulate}}.}
}
\value{
An ECMMon object.
}
\description{
Batch simulates over a data frame of parameters.
}
\details{
If the parameters argument \code{params} is a list of numerical vectors
the batch simulation parameters data frame is generated with \code{\link{expand.grid}}.
}
\seealso{
Other Simulation functions: 
\code{\link{ECMMonAssignInitialConditions}()},
\code{\link{ECMMonAssignRateValues}()},
\code{\link{ECMMonExtendByAdjacencyMatrix}()},
\code{\link{ECMMonSimulate}()}
}
\concept{Simulation functions}
