% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterfaceFunctions.R
\name{ECMMonCreateDasboard}
\alias{ECMMonCreateDasboard}
\title{Creation of an ECMMon dashboard.}
\usage{
ECMMonCreateDasboard(ecmObj, dashboardTheme = NULL)
}
\arguments{
\item{ecmObj}{An ECMMon object.}

\item{dashboardTheme}{A string specifying the dashboard theme.
See \code{\link{dashboardthemes::shinyDashboardThemes}}.
Ignored if the package \code{dashboardthemes} is not installed.}
}
\value{
Shiny app
}
\description{
Creation of an ECMMon dashboard.
}
\seealso{
Other ECMMon interface functions: 
\code{\link{ECMMonDashboardServerFunction}()},
\code{\link{ECMMonDashboardUI}()}
}
\concept{ECMMon interface functions}
