% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterfaceFunctions.R
\name{ECMMonDashboardServerFunction}
\alias{ECMMonDashboardServerFunction}
\title{ECMMon dashboard server function}
\usage{
ECMMonDashboardServerFunction(ecmObj)
}
\arguments{
\item{ecmObj}{A time series recommender.}
}
\value{
Shiny server function.
}
\description{
Creates the Shiny server function for a sparse matrix recommender interface.
}
\seealso{
Other ECMMon interface functions: 
\code{\link{ECMMonCreateDasboard}()},
\code{\link{ECMMonDashboardUI}()}
}
\concept{ECMMon interface functions}
