% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterfaceFunctions.R
\name{ECMMonDashboardUI}
\alias{ECMMonDashboardUI}
\title{ECMMon dashboard UI}
\usage{
ECMMonDashboardUI(ecmObj, dashboardTheme = NULL)
}
\arguments{
\item{ecmObj}{A ECMMon object.}

\item{dashboardTheme}{A string specifying the dashboard theme.
See \code{\link{dashboardthemes::shinyDashboardThemes}}.
Ignored if the package \code{dashboardthemes} is not installed.}
}
\value{
Shiny UI object.
}
\description{
Creates the Shiny UI function for an ECMMon interface.
}
\seealso{
Other ECMMon interface functions: 
\code{\link{ECMMonCreateDasboard}()},
\code{\link{ECMMonDashboardServerFunction}()}
}
\concept{ECMMon interface functions}
