% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECMMon.R
\name{ECMMonEcho}
\alias{ECMMonEcho}
\title{Echo.}
\usage{
ECMMonEcho(ecmObj, x)
}
\arguments{
\item{ecmObj}{An ECMMon object.}

\item{x}{Object to be echoed.}
}
\value{
An ECMMon object
}
\description{
Echoes the given argument.
}
\concept{Echo functions}
