% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECMMon.R
\name{ECMMonEchoFunctionValue}
\alias{ECMMonEchoFunctionValue}
\title{Echo function application to monad's value.}
\usage{
ECMMonEchoFunctionValue(ecmObj, f)
}
\arguments{
\item{ecmObj}{An ECMMon object.}

\item{f}{A function to be applied to \code{ecmObj$Value}.}
}
\value{
A ECMMon object.
}
\description{
Applies a function to the "Value" element/member of the monad object
and prints the result.
}
\details{
Prints \code{f(ecmObj$Value)}.
}
