% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECMMon.R
\name{ECMMonEchoValue}
\alias{ECMMonEchoValue}
\title{Echo monad's value.}
\usage{
ECMMonEchoValue(ecmObj)
}
\arguments{
\item{ecmObj}{An ECMMon object.}
}
\value{
A ECMMon object.
}
\description{
Prints the "Value" element/member of the monad object.
}
\details{
Prints \code{ecmObj$Value}.
}
