% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECMMon.R
\name{ECMMonExtendByAdjacencyMatrix}
\alias{ECMMonExtendByAdjacencyMatrix}
\title{Extend by adjacency matrix}
\usage{
ECMMonExtendByAdjacencyMatrix(ecmObj, mat, migratingStocks = NULL, ...)
}
\arguments{
\item{ecmObj}{An ECMMon object.}

\item{mat}{A matrix.}

\item{migratingStocks}{A a character vector with stocks (names.)}

\item{...}{Additional arguments for \code{ToSiteCompartmentalModel}.}
}
\value{
An ECMMon object.
}
\description{
Extends monad's single site model into multi-site model using the numerical matrix mat.
}
\seealso{
Other Simulation functions: 
\code{\link{ECMMonAddRates}()},
\code{\link{ECMMonAssignInitialConditions}()},
\code{\link{ECMMonAssignRateValues}()},
\code{\link{ECMMonBatchSimulate}()},
\code{\link{ECMMonSimulate}()}
}
\concept{Simulation functions}
