% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECMMon.R
\docType{data}
\name{ECMMonFailureSymbol}
\alias{ECMMonFailureSymbol}
\title{Failure symbol for ECMMon.}
\format{
An object of class \code{logical} of length 1.
}
\usage{
ECMMonFailureSymbol
}
\description{
Failure symbol for the monad ECMMon.
}
\keyword{datasets}
