% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECMMon.R
\name{ECMMonGetSolutionLongForm}
\alias{ECMMonGetSolutionLongForm}
\title{Get long form solution.}
\usage{
ECMMonGetSolutionLongForm(
  ecmObj,
  stockDescriptionsQ = TRUE,
  siteIdentifiersQ = TRUE
)
}
\arguments{
\item{ecmObj}{An ECMMon object.}

\item{stockDescriptionsQ}{Should a column with the stock descriptions be added or not?}

\item{siteIdentifiersQ}{Should a column with site identifiers be added or not?
If \code{ecmObj} has only a single-site model, the that column will have all zeroes.}
}
\value{
An ECMMon object.
}
\description{
Gets the long form of the solution data frame \code{ecmObj$Solution}.
}
\details{
The stocks types are derived from the stocks descriptions, \code{ecmObj$Stocks}.
The stocks types are "Population", "Money", "Other".
}
\seealso{
Other Get functions: 
\code{\link{ECMMonGetDefaultModel}()}
}
\concept{Get functions}
