% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterfaceFunctions.R
\name{ECMMonMakeSearchUI}
\alias{ECMMonMakeSearchUI}
\title{ECMMon interface UI}
\usage{
ECMMonMakeSearchUI(smrObj, dashboardTheme = NULL)
}
\arguments{
\item{dashboardTheme}{A string specifying the dashboard theme.
See \code{\link{dashboardthemes::shinyDashboardThemes}}.
Ignored if the package \code{dashboardthemes} is not installed.}

\item{ecmObj}{A ECMMon object.}
}
\value{
Shiny UI object.
}
\description{
Creates the Shiny UI function for an ECMMon interface.
}
\concept{ECMMon interface functions}
