% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECMMon.R
\name{ECMMonOption}
\alias{ECMMonOption}
\title{Optional function application to monad's object.}
\usage{
ECMMonOption(ecmObj, f)
}
\arguments{
\item{ecmObj}{An ECMMon object.}

\item{f}{A function to be applied to the monad object.}
}
\value{
A ECMMon object.
}
\description{
If monadic failure is obtained from \code{f(ecmObj)}
then returns the original \code{ecmObj};
else returns the result of \code{f(ecmObj)}.
}
\details{
In general \code{f} should return a monad object,
but that is not enforced.
}
