% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECMMon.R
\name{ECMMonPlotSolutionHistograms}
\alias{ECMMonPlotSolutionHistograms}
\title{Plot solution histograms.}
\usage{
ECMMonPlotSolutionHistograms(
  ecmObj,
  stocksSpec = NULL,
  maxTime = NULL,
  echoQ = TRUE,
  stockDescriptionsQ = TRUE,
  scales = "free",
  ncol = 2,
  ...
)
}
\arguments{
\item{ecmObj}{An ECMMon object.}

\item{stocksSpec}{A character with stock names or string patterns for stock names to be plotted.}

\item{maxTime}{A numerical non-negative value for the maximum simulation time or NULL.
If NULL the maximum time is derived from \code{ecmObj$Solution}.}

\item{echoQ}{Should the plot be echoed or not?}

\item{stockDescriptionsQ}{Should the stock descriptions be added or not?}

\item{scales}{Scales argument \code{scales} for \code{\link{ggplot2::facet_wrap}}.}

\item{ncol}{Number of columns \code{ncol} argument for \code{\link{ggplot2::facet_wrap}}.}

\item{...}{Additional parameters for \code{\link{ggplot2::geom_histogram}}.}
}
\value{
An ECMMon object.
}
\description{
Plots solution histograms.
}
\details{
The plot is assigned to \code{ecmObj$Value}.
}
\seealso{
Other Plot functions: 
\code{\link{ECMMonPlotSolutions}()}
}
\concept{Plot functions}
