% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECMMon.R
\name{ECMMonPlotSolutions}
\alias{ECMMonPlotSolutions}
\title{Plot solutions.}
\usage{
ECMMonPlotSolutions(ecmObj, stocksSpec = NULL, maxTime = NULL, echoQ = TRUE)
}
\arguments{
\item{ecmObj}{An ECMMon object.}

\item{stocksSpec}{A character with stock names or string patterns for stock names to be plotted.}

\item{maxTime}{A numerical non-negative value for the maximum simulation time or NULL.
If NULL the maximum time is derived from \code{ecmObj$Solution}.}

\item{echoQ}{Should the plot be echoed or not?}
}
\value{
An ECMMon object.
}
\description{
Plots solutions.
}
\details{
The plot is assigned to \code{ecmObj$Value}.
}
\concept{Plot functions}
