% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECMMon.R
\name{ECMMonSetGrid}
\alias{ECMMonSetGrid}
\title{Set Grid.}
\usage{
ECMMonSetGrid(ecmObj, Grid)
}
\arguments{
\item{ecmObj}{An ECMMon object.}

\item{Grid}{An object member to be set.}
}
\value{
An ECMMon object.
}
\description{
Sets Grid into the monad object.
}
\seealso{
Other Set/Take functions: 
\code{\link{ECMMonSetDeSolveSolution}()},
\code{\link{ECMMonSetMultiSiteModel}()},
\code{\link{ECMMonSetSingleSiteModel}()},
\code{\link{ECMMonSetSolution}()},
\code{\link{ECMMonSetValue}()},
\code{\link{ECMMonTakeDeSolveSolution}()},
\code{\link{ECMMonTakeGrid}()},
\code{\link{ECMMonTakeMultiSiteModel}()},
\code{\link{ECMMonTakeSingleSiteModel}()},
\code{\link{ECMMonTakeSolution}()},
\code{\link{ECMMonTakeValue}()}
}
\concept{Set/Take functions}
