% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECMMon.R
\name{ECMMonSetMultiSiteModel}
\alias{ECMMonSetMultiSiteModel}
\title{Set MultiSiteModel.}
\usage{
ECMMonSetMultiSiteModel(ecmObj, MultiSiteModel)
}
\arguments{
\item{ecmObj}{An ECMMon object.}

\item{MultiSiteModel}{An object member to be set.}
}
\value{
An ECMMon object.
}
\description{
Sets MultiSiteModel into the monad object.
}
\seealso{
Other Set/Take functions: 
\code{\link{ECMMonSetDeSolveSolution}()},
\code{\link{ECMMonSetGrid}()},
\code{\link{ECMMonSetSingleSiteModel}()},
\code{\link{ECMMonSetSolution}()},
\code{\link{ECMMonSetValue}()},
\code{\link{ECMMonTakeDeSolveSolution}()},
\code{\link{ECMMonTakeGrid}()},
\code{\link{ECMMonTakeMultiSiteModel}()},
\code{\link{ECMMonTakeSingleSiteModel}()},
\code{\link{ECMMonTakeSolution}()},
\code{\link{ECMMonTakeValue}()}
}
\concept{Set/Take functions}
