% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECMMon.R
\name{ECMMonSetSingleSiteModel}
\alias{ECMMonSetSingleSiteModel}
\title{Set SingleSiteModel.}
\usage{
ECMMonSetSingleSiteModel(ecmObj, SingleSiteModel)
}
\arguments{
\item{ecmObj}{An ECMMon object.}

\item{SingleSiteModel}{An object member to be set.}
}
\value{
An ECMMon object.
}
\description{
Sets SingleSiteModel into the monad object.
}
\seealso{
Other Set/Take functions: 
\code{\link{ECMMonSetDeSolveSolution}()},
\code{\link{ECMMonSetGrid}()},
\code{\link{ECMMonSetMultiSiteModel}()},
\code{\link{ECMMonSetSolution}()},
\code{\link{ECMMonSetValue}()},
\code{\link{ECMMonTakeDeSolveSolution}()},
\code{\link{ECMMonTakeGrid}()},
\code{\link{ECMMonTakeMultiSiteModel}()},
\code{\link{ECMMonTakeSingleSiteModel}()},
\code{\link{ECMMonTakeSolution}()},
\code{\link{ECMMonTakeValue}()}
}
\concept{Set/Take functions}
