% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECMMon.R
\name{ECMMonSetSolution}
\alias{ECMMonSetSolution}
\title{Set Solution.}
\usage{
ECMMonSetSolution(ecmObj, Solution)
}
\arguments{
\item{ecmObj}{An ECMMon object.}

\item{Solution}{An object member to be set.}
}
\value{
An ECMMon object.
}
\description{
Sets Solution into the monad object.
}
\seealso{
Other Set/Take functions: 
\code{\link{ECMMonSetGrid}()},
\code{\link{ECMMonSetMultiSiteModel}()},
\code{\link{ECMMonSetSingleSiteModel}()},
\code{\link{ECMMonSetValue}()},
\code{\link{ECMMonSetdeSolveSolution}()},
\code{\link{ECMMonTakeGrid}()},
\code{\link{ECMMonTakeMultiSiteModel}()},
\code{\link{ECMMonTakeSingleSiteModel}()},
\code{\link{ECMMonTakeSolution}()},
\code{\link{ECMMonTakeValue}()},
\code{\link{ECMMonTakedeSolveSolution}()}
}
\concept{Set/Take functions}
