% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECMMon.R
\name{ECMMonSetValue}
\alias{ECMMonSetValue}
\title{Set the value in a ECMMon object.}
\usage{
ECMMonSetValue(ecmObj, value)
}
\arguments{
\item{ecmObj}{An ECMMon object.}

\item{value}{The new value.}
}
\value{
A ECMMon object.
}
\description{
Sets the value in a ECMMon monad object.
}
\details{
Assigns \code{value} to \code{ecmObj$Value}.
}
\seealso{
Other Set/Take functions: 
\code{\link{ECMMonSetDeSolveSolution}()},
\code{\link{ECMMonSetGrid}()},
\code{\link{ECMMonSetMultiSiteModel}()},
\code{\link{ECMMonSetSingleSiteModel}()},
\code{\link{ECMMonSetSolution}()},
\code{\link{ECMMonTakeDeSolveSolution}()},
\code{\link{ECMMonTakeGrid}()},
\code{\link{ECMMonTakeMultiSiteModel}()},
\code{\link{ECMMonTakeSingleSiteModel}()},
\code{\link{ECMMonTakeSolution}()},
\code{\link{ECMMonTakeValue}()}
}
\concept{Set/Take functions}
