% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECMMon.R
\name{ECMMonSimulate}
\alias{ECMMonSimulate}
\title{Simulate.}
\usage{
ECMMonSimulate(ecmObj, maxTime, ...)
}
\arguments{
\item{ecmObj}{An ECMMon object.}

\item{maxTime}{A numerical non-negative value for the maximum simulation time.}

\item{...}{Additional parameters of \code{\link{deSolve::ode}}.}
}
\value{
An ECMMon object.
}
\description{
If there is a multi-site models simulates with that model for the specified maximum time.
Otherwise the simulation is done with single-site model.
}
\seealso{
Other Simulation functions: 
\code{\link{ECMMonAssignInitialConditions}()},
\code{\link{ECMMonAssignRateValues}()},
\code{\link{ECMMonBatchSimulate}()},
\code{\link{ECMMonExtendByAdjacencyMatrix}()}
}
\concept{Simulation functions}
