% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECMMon.R
\name{ECMMonTakedeSolveSolution}
\alias{ECMMonTakedeSolveSolution}
\title{Take deSolveSolution.}
\usage{
ECMMonTakedeSolveSolution(ecmObj, functionName = "ECMMonTakedeSolveSolution")
}
\arguments{
\item{ecmObj}{An ECMMon object.}

\item{functionName}{A string that is a name of this function or a delegating function.}
}
\value{
A list of functions or \code{ECMMonFailureSymbol}.
}
\description{
Takes deSolveSolution from the monad object.
}
\seealso{
Other Set/Take functions: 
\code{\link{ECMMonSetGrid}()},
\code{\link{ECMMonSetMultiSiteModel}()},
\code{\link{ECMMonSetSingleSiteModel}()},
\code{\link{ECMMonSetSolution}()},
\code{\link{ECMMonSetValue}()},
\code{\link{ECMMonSetdeSolveSolution}()},
\code{\link{ECMMonTakeGrid}()},
\code{\link{ECMMonTakeMultiSiteModel}()},
\code{\link{ECMMonTakeSingleSiteModel}()},
\code{\link{ECMMonTakeSolution}()},
\code{\link{ECMMonTakeValue}()}
}
\concept{Set/Take functions}
