% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECMMon.R
\name{ECMMonTakeGrid}
\alias{ECMMonTakeGrid}
\title{Take Grid.}
\usage{
ECMMonTakeGrid(ecmObj, functionName = "ECMMonTakeGrid")
}
\arguments{
\item{ecmObj}{An ECMMon object.}

\item{functionName}{A string that is a name of this function or a delegating function.}
}
\value{
A list of functions or \code{ECMMonFailureSymbol}.
}
\description{
Takes Grid from the monad object.
}
\seealso{
Other Set/Take functions: 
\code{\link{ECMMonSetGrid}()},
\code{\link{ECMMonSetMultiSiteModel}()},
\code{\link{ECMMonSetSingleSiteModel}()},
\code{\link{ECMMonSetSolution}()},
\code{\link{ECMMonSetValue}()},
\code{\link{ECMMonSetdeSolveSolution}()},
\code{\link{ECMMonTakeMultiSiteModel}()},
\code{\link{ECMMonTakeSingleSiteModel}()},
\code{\link{ECMMonTakeSolution}()},
\code{\link{ECMMonTakeValue}()},
\code{\link{ECMMonTakedeSolveSolution}()}
}
\concept{Set/Take functions}
