% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECMMon.R
\name{ECMMonTakeMultiSiteModel}
\alias{ECMMonTakeMultiSiteModel}
\title{Take MultiSiteModel.}
\usage{
ECMMonTakeMultiSiteModel(ecmObj, functionName = "ECMMonTakeMultiSiteModel")
}
\arguments{
\item{ecmObj}{An ECMMon object.}

\item{functionName}{A string that is a name of this function or a delegating function.}
}
\value{
A list of functions or \code{ECMMonFailureSymbol}.
}
\description{
Takes MultiSiteModel from the monad object.
}
\seealso{
Other Set/Take functions: 
\code{\link{ECMMonSetDeSolveSolution}()},
\code{\link{ECMMonSetGrid}()},
\code{\link{ECMMonSetMultiSiteModel}()},
\code{\link{ECMMonSetSingleSiteModel}()},
\code{\link{ECMMonSetSolution}()},
\code{\link{ECMMonSetValue}()},
\code{\link{ECMMonTakeDeSolveSolution}()},
\code{\link{ECMMonTakeGrid}()},
\code{\link{ECMMonTakeSingleSiteModel}()},
\code{\link{ECMMonTakeSolution}()},
\code{\link{ECMMonTakeValue}()}
}
\concept{Set/Take functions}
