% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECMMon.R
\name{ECMMonTakeValue}
\alias{ECMMonTakeValue}
\title{Take the value in a ECMMon object.}
\usage{
ECMMonTakeValue(ecmObj)
}
\arguments{
\item{ecmObj}{An ECMMon object.}
}
\value{
Just \code{ecmObj$Value}.
}
\description{
Takes the value from ECMMon monad object.
}
\seealso{
Other Set/Take functions: 
\code{\link{ECMMonSetDeSolveSolution}()},
\code{\link{ECMMonSetGrid}()},
\code{\link{ECMMonSetMultiSiteModel}()},
\code{\link{ECMMonSetSingleSiteModel}()},
\code{\link{ECMMonSetSolution}()},
\code{\link{ECMMonSetValue}()},
\code{\link{ECMMonTakeDeSolveSolution}()},
\code{\link{ECMMonTakeGrid}()},
\code{\link{ECMMonTakeMultiSiteModel}()},
\code{\link{ECMMonTakeSingleSiteModel}()},
\code{\link{ECMMonTakeSolution}()}
}
\concept{Set/Take functions}
