% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECMMon.R
\name{ECMMonUnit}
\alias{ECMMonUnit}
\title{Make a ECMMon Unit}
\usage{
ECMMonUnit(model = NULL)
}
\arguments{
\item{model}{A single site model.}
}
\value{
An S3 class "ECMMon". In other words, a list with the attribute "class" set to "ECMMon".
}
\description{
Creates a monad object.
}
