% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECMMon.R
\name{ECMMonVerifyDataDirectory}
\alias{ECMMonVerifyDataDirectory}
\title{Verify does a directory have ECMMon data files.}
\usage{
ECMMonVerifyDataDirectory(directoryName)
}
\arguments{
\item{directoryName}{A directory name string.}
}
\value{
A logical value.
}
\description{
Verify does a directory have the CSV files \code{model.csv}
and \code{solution.csv}.
}
\concept{Non-monadic functions.}
