% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpidemiologyModels.R
\name{EpidemiologyFullModelProperlyDefinedQ}
\alias{EpidemiologyFullModelProperlyDefinedQ}
\title{Properly defined epidemiology full model object check.}
\usage{
EpidemiologyFullModelProperlyDefinedQ(model)
}
\arguments{
\item{model}{An object}
}
\value{
A logical.
}
\description{
Checks is an object a full epidemiology model object and
are its rates and stocks are properly defined.
}
\seealso{
Other Epidemiology Modeling functions: 
\code{\link{EpidemiologyFullModelQ}()},
\code{\link{EpidemiologyModelQ}()}
}
\concept{Epidemiology Modeling functions}
