% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpidemiologyModels.R
\name{EpidemiologyFullModelQ}
\alias{EpidemiologyFullModelQ}
\title{Epidemiology full model object check.}
\usage{
EpidemiologyFullModelQ(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical.
}
\description{
Checks is an object an epidemiology model object.
}
\seealso{
Other Epidemiology Modeling functions: 
\code{\link{EpidemiologyFullModelProperlyDefinedQ}()},
\code{\link{EpidemiologyModelQ}()}
}
\concept{Epidemiology Modeling functions}
