% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpidemiologyModels.R
\name{EpidemiologyModelQ}
\alias{EpidemiologyModelQ}
\title{Epidemiology model object check.}
\usage{
EpidemiologyModelQ(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A logical.
}
\description{
Checks is an object an epidemiology model object.
}
\seealso{
Other Epidemiology Modeling functions: 
\code{\link{EpidemiologyFullModelProperlyDefinedQ}()},
\code{\link{EpidemiologyFullModelQ}()}
}
\concept{Epidemiology Modeling functions}
