% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpidemiologyModelModifications.R
\name{GetStocks}
\alias{GetStocks}
\title{Get stock symbols.}
\usage{
GetStocks(model, patterns, applyToStockNamesQ = FALSE)
}
\arguments{
\item{model}{A model object.}

\item{patterns}{A a character vector of patterns to be applied to the descriptions of model's stocks.}

\item{applyToStockNamesQ}{Should the patterns be applied also to the stock names or not?}
}
\value{
A character vector
}
\description{
Gets the stock symbols that have descriptions (or names) matching specified patterns.
}
\seealso{
Other Epidemiology Model Query functions: 
\code{\link{GetRates}()}
}
\concept{Epidemiology Model Query functions}
