% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpidemiologyModelModifications.R
\name{JoinModels}
\alias{JoinModels}
\title{Join two models.}
\usage{
JoinModels(model1, model2)
}
\arguments{
\item{model1}{A model object.}

\item{model2}{A model object.}
}
\value{
A model object.
}
\description{
Makes a multi-site model with the stocks, rates, and equations
of arguments joined.
}
\seealso{
Other Epidemiology Model Modifier functions: 
\code{\link{AddModelIdentifier}()},
\code{\link{AddModelRates}()},
\code{\link{MakeMigrationEquations}()},
\code{\link{TakeRHSFunctionBodyLines}()},
\code{\link{ToSiteCompartmentalModel}()}
}
\concept{Epidemiology Model Modifier functions}
