% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpidemiologyModelModifications.R
\name{MakeCoreMultiSiteModel}
\alias{MakeCoreMultiSiteModel}
\title{Make core multi-site model.}
\usage{
MakeCoreMultiSiteModel(model, cellIdentifiers)
}
\arguments{
\item{model}{A model object.}

\item{cellIdentifiers}{A a character vector of cell identifiers.}
}
\value{
A model object
}
\description{
Makes core multi-site model.
}
\concept{Epidemiology Model Modification functions}
