% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpidemiologyModelModifications.R
\name{MakeMigrationEquations}
\alias{MakeMigrationEquations}
\title{Make migration equations.}
\usage{
MakeMigrationEquations(
  data,
  migratingStocks,
  totalStock = "TPt",
  nLeadingWhitespace = 8
)
}
\arguments{
\item{data}{A matrix or a data frame with columns \code{c('i', 'j', 'x')}.}

\item{migratingStocks}{A a character vector with stocks.}

\item{totalStock}{A string}

\item{nLeadingWhitespace}{Number of leading whitespace characters for each equation.}
}
\value{
A character vector
}
\description{
Makes the migration (supplemental) equations with the specified matrix and stocks.
}
\seealso{
Other Epidemiology Model Modifier functions: 
\code{\link{AddModelIdentifier}()},
\code{\link{AddModelRates}()},
\code{\link{JoinModels}()},
\code{\link{TakeRHSFunctionBodyLines}()},
\code{\link{ToSiteCompartmentalModel}()}
}
\concept{Epidemiology Model Modifier functions}
