% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpidemiologyModelModifications.R
\name{SparseMatrixQ}
\alias{SparseMatrixQ}
\title{Sparse matrix check}
\usage{
SparseMatrixQ(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A logical value
}
\description{
Sparse matrix check
}
