% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpidemiologyModelModifications.R
\name{ToSiteCompartmentalModel}
\alias{ToSiteCompartmentalModel}
\title{Crate site compartmental model}
\usage{
ToSiteCompartmentalModel(
  model,
  mat,
  migratingStocks = NULL,
  cellIdentifiers = NULL,
  ...
)
}
\arguments{
\item{model}{A (single-site) model object.}

\item{mat}{A matrix.}

\item{migratingStocks}{A a character vector with stocks (names.)}

\item{cellIdentifiers}{A character vector with cell identifiers.
If NULL then \code{unique(c(rownames(mat), colnames(mat)))} is used.}

\item{...}{Additional arguments for \code{MakeMigrationEquations}.}
}
\value{
A character vector
}
\description{
Makes the migration (supplemental) equations with the specified matrix and stocks.
}
\seealso{
Other Epidemiology Model Modifier functions: 
\code{\link{AddModelIdentifier}()},
\code{\link{AddModelRates}()},
\code{\link{JoinModels}()},
\code{\link{MakeMigrationEquations}()},
\code{\link{TakeRHSFunctionBodyLines}()}
}
\concept{Epidemiology Model Modifier functions}
