% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpidemiologyModelModifications.R
\name{AddModelIdentifier}
\alias{AddModelIdentifier}
\title{Add a model identifier.}
\usage{
AddModelIdentifier(
  model,
  id,
  addToOtherLeftHandSidesQ = TRUE,
  birthsTermQ = FALSE
)
}
\arguments{
\item{model}{A model object.}

\item{id}{A number or string to used as ID.}

\item{addToOtherLeftHandSidesQ}{Should the id added to other left hand side symbols or not?}

\item{birthsTermQ}{Should the births term be included or not?}
}
\value{
A model.
}
\description{
Adds a specified identifier id to all stocks and rates on the model m.
}
\details{
The original model object has \code{birthTermsQ} in the closure of \code{model$RHSFunction}.
(For now) it is easier to just have that parameter specified than imply it.
}
\seealso{
Other Epidemiology Model Modifier functions: 
\code{\link{AddModelRates}()},
\code{\link{JoinModels}()},
\code{\link{MakeMigrationEquations}()},
\code{\link{TakeRHSFunctionBodyLines}()},
\code{\link{ToSiteCompartmentalModel}()}
}
\concept{Epidemiology Model Modifier functions}
