% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECMMon.R
\name{ECMMonGetDefaultModel}
\alias{ECMMonGetDefaultModel}
\title{Get default model object.}
\usage{
ECMMonGetDefaultModel(ecmObj)
}
\arguments{
\item{ecmObj}{An ECMMon object.}
}
\value{
An ECMMon object.
}
\description{
Gets the default model of the monad object.
}
\seealso{
Other Get functions: 
\code{\link{ECMMonGetSolutionLongForm}()}
}
\concept{Get functions}
