% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECMMon.R
\name{ECMMonPlotSolutions}
\alias{ECMMonPlotSolutions}
\title{Plot solutions.}
\usage{
ECMMonPlotSolutions(
  ecmObj,
  stocksSpec = NULL,
  maxTime = NULL,
  echoQ = TRUE,
  stockDescriptionsQ = TRUE,
  separatePlotsQ = FALSE,
  ...
)
}
\arguments{
\item{ecmObj}{An ECMMon object.}

\item{stocksSpec}{A character with stock names or string patterns for stock names to be plotted.}

\item{maxTime}{A numerical non-negative value for the maximum simulation time or NULL.
If NULL the maximum time is derived from \code{ecmObj$Solution}.}

\item{echoQ}{Should the plot be echoed or not?}

\item{stockDescriptionsQ}{Should the stock descriptions be added or not?}

\item{separatePlotsQ}{Should the stock plots be separated or not?}

\item{...}{Additional parameters for \code{\link{ggplot2::facet_wrap}}}
}
\value{
An ECMMon object.
}
\description{
Plots solutions.
}
\details{
The plot is assigned to \code{ecmObj$Value}.
}
\seealso{
Other Plot functions: 
\code{\link{ECMMonPlotSolutionHistograms}()}
}
\concept{Plot functions}
