% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECMMon.R
\name{ECMMonSetdeSolveSolution}
\alias{ECMMonSetdeSolveSolution}
\title{Set deSolveSolution.}
\usage{
ECMMonSetdeSolveSolution(ecmObj, deSolveSolution)
}
\arguments{
\item{ecmObj}{An ECMMon object.}

\item{deSolveSolution}{An object member to be set.}
}
\value{
An ECMMon object.
}
\description{
Sets deSolveSolution into the monad object.
}
\seealso{
Other Set/Take functions: 
\code{\link{ECMMonSetGrid}()},
\code{\link{ECMMonSetMultiSiteModel}()},
\code{\link{ECMMonSetSingleSiteModel}()},
\code{\link{ECMMonSetSolution}()},
\code{\link{ECMMonSetValue}()},
\code{\link{ECMMonTakeGrid}()},
\code{\link{ECMMonTakeMultiSiteModel}()},
\code{\link{ECMMonTakeSingleSiteModel}()},
\code{\link{ECMMonTakeSolution}()},
\code{\link{ECMMonTakeValue}()},
\code{\link{ECMMonTakedeSolveSolution}()}
}
\concept{Set/Take functions}
