% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpidemiologyModelModifications.R
\name{GetRates}
\alias{GetRates}
\title{Get rate symbols.}
\usage{
GetRates(model, patterns)
}
\arguments{
\item{model}{A model object.}

\item{patterns}{A a character vector of patterns to be applied to the descriptions of model's rates.}
}
\value{
A character vector
}
\description{
Gets the rate symbols that have descriptions matching specified patterns.
}
\seealso{
Other Epidemiology Model Query functions: 
\code{\link{GetStocks}()},
\code{\link{MakeCoreMultiSiteModel}()}
}
\concept{Epidemiology Model Query functions}
