% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpidemiologyModelModifications.R
\name{TakeRHSFunctionBodyLines}
\alias{TakeRHSFunctionBodyLines}
\title{Take RHSFunction body lines.}
\usage{
TakeRHSFunctionBodyLines(model)
}
\arguments{
\item{model}{A model object.}
}
\value{
A character vector
}
\description{
Takes RHSFunction body lines of a model.
}
\seealso{
Other Epidemiology Model Modifier functions: 
\code{\link{AddModelIdentifier}()},
\code{\link{AddModelRates}()},
\code{\link{JoinModels}()},
\code{\link{MakeMigrationEquations}()},
\code{\link{ToSiteCompartmentalModel}()}
}
\concept{Epidemiology Model Modifier functions}
